requirements.txt
clickhouse-driver==0.2.9
mysql-connector-python==8.1.0
PyYAML==6.0
